requirements.txt
asgiref==3.6.0
Django==4.0.10
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2023.3
sqlparse==0.4.4
whitenoise==6.0.0
